ajute id

+++ SmartDigitalPsico/SmartDigitalPsico.Data/Contract/IAuthRepository.cs
@@ -6,7 +6,7 @@ namespace SmartDigitalPsico.Data.Contract
 {
     public interface IAuthRepository
     {
-        Task<ServiceResponse<int>> Register(User user, string password);
+        Task<ServiceResponse<long>> Register(User user, string password);
         Task<ServiceResponse<string>> Login(string username, string password);
         Task<bool> UserExists(string username);
     }
